IBM-Swift/Kitura#888 Improve user profile for Google token
